로그아웃 기능 어느 페이지에서든 구현가능하게 수정
