+++ src/main/java/momentlockdemo/controller/GlobalControllerAdvice.java
@@ -1,0 +1,17 @@
+package momentlockdemo.controller;
+
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ModelAttribute;
+
+@ControllerAdvice
+public class GlobalControllerAdvice {
+
+	@ModelAttribute
+	public void addUserToModel(Model model, @AuthenticationPrincipal User user) {
+
+		model.addAttribute("user", user);
+	}
+}
