+++ src/main/resources/static/js/main.js
@@ -1,40 +1,3 @@
-const loginModalButton = document.getElementById("modalLoginBtn");
-
-
-async function logout() {
-
-	try {
-
-		const response = await fetch("/logout", {
-
-			mehtod: 'POST',
-			headders: {
-
-				'Content-Type': 'application/JSON'
-
-			}
-
-		});
-
-		if (response.ok) {
-
-			window.location.href = '/html/main';
-
-		} else {
-
-			alert("logout failed");
-
-		}
-
-	} catch (error) {
-
-		alert("error occured during logout");
-
-	}
-
-};
-
-
 document.addEventListener("DOMContentLoaded", () => {
 	const startBtn = document.getElementById("startCapsuleBtn");
 	if (startBtn) {
@@ -63,8 +26,4 @@ document.addEventListener("DOMContentLoaded", () => {
 });
 
 
-function gotoLogin() {
 
-	window.location.href = '/momentlock/login';
-
-};
